pyproject.toml
[tool.poetry]
name = "hw19-2"
version = "0.1.0"
description = ""
authors = ["Alex Urapin <kulob1@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.16"
psycopg2 = "^2.9"
pillow = "^10.4.0"
ipython = "^8.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.12.50"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
